requirements.txt
requests~=2.32.3
aiogram~=3.17.0
python-dotenv~=1.0.1
aiohttp~=3.11.12